day 7

+++ samples/mnist_mlp.py
@@ -33,8 +33,8 @@ x_test = x_test.reshape(x_test.shape[0], img_rows*img_cols)
 input_shape = (img_rows*img_cols,)
 
 # When calculating image data, convert from uint8 to float
-x_train = x_train.astype('float32')
-x_test = x_test.astype('float32')
+x_train = x_train.astype('float')
+x_test = x_test.astype('float')
 
 # Reduce the element range from [0, 255] to [0, 1]
 x_train /= 255
